requirements.txt
# Core dependencies
torch>=1.9.0
transformers>=4.20.0
opencv-python>=4.5.0
whisper-openai>=20231117
moviepy>=1.0.3
numpy>=1.21.0
pandas>=1.3.0

# NLP and AI models
sentence-transformers>=2.2.0
nltk>=3.7

# Web framework
streamlit>=1.28.0
flask>=2.3.0

# Video processing
ffmpeg-python>=0.2.0
Pillow>=9.0.0

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
tqdm>=4.64.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
